src/requirements.txt
streamlit==1.26.0
st-pages==0.4.4
Pillow==9.0.1
streamlit-extras==0.3.0
